+++ gaios/Home/Controllers/HomeViewController.swift
@@ -109,23 +109,29 @@ class HomeViewController: UIViewController {
             return nil
         }
     }
-    
-    func onTap(_ indexPath: IndexPath) {
-        if let account = getAccountFromTableView(indexPath) {
+
+    func goAccount(account: Account) {
+        if let wm = WalletsRepository.shared.get(for: account.id), wm.logged {
+            AccountNavigator.goLogged(account: account)
+        } else {
             AccountNavigator.goLogin(account: account)
         }
+    }
+
+    func onTap(_ indexPath: IndexPath) {
+        onTapOverview(indexPath)
     }
     
     func onTapOverview(_ indexPath: IndexPath) {
         if let account = getAccountFromTableView(indexPath) {
-            AccountNavigator.goLogin(account: account)
+            goAccount(account: account)
         }
     }
     
     func onTapLightShort(_ indexPath: IndexPath) {
         if let account = getAccountFromTableView(indexPath) {
             if let lightning = account.getDerivedLightningAccount() {
-                AccountNavigator.goLogin(account: lightning)
+                goAccount(account: lightning)
             }
         }
     }
@@ -324,7 +330,6 @@ extension HomeViewController: UITableViewDelegate, UITableViewDataSource {
     }
 
     func tableView(_ tableView: UITableView, didSelectRowAt indexPath: IndexPath) {
-        
     }
 }