+++ gaios/Wallet/AccountViewController/AccountViewModel.swift
@@ -11,6 +11,7 @@ enum AmpEducationalMode {
 
 class AccountViewModel {
     var wm: WalletManager? { WalletManager.current }
+    var active: Bool { wm?.activeSessions.allSatisfy { !$0.value.paused } ?? false }
     var cachedBalance: AssetAmountList
     var cachedTransactions = [Transaction]()
     var account: WalletItem!